Introduce `whitelisted_authors` and `whitelisted_authors_regex` options
If the `whitelisted_authors_regex` is provided the author email will be
checked with this value. If it matches author will be notifiable if not
and `whitelisted_authors` is provided, the email of the author will be
tested with the given list. If the email of the author is in this list
author will be notifiable, otherwise marked as not notifiable.

Long story short; `whitelisted_authors` always override the regex option
if both of them are provided. Otherwise the logic is quite simple check,
as their name suggest.